+++ task/webpack.example.config.js
@@ -65,7 +65,11 @@ module.exports = {
           MiniCssExtractPlugin.loader,
           'css-loader',
         ]
-      }
+      },
+      {
+        test: /\.svg$/,
+        loader: 'svg-inline-loader?classPrefix',
+      },
     ],
   },
   devtool: 'source-map',